Main : Saperate event listener
